📝 어스 리팩토링
어스 책임 분리 진행
