+++ schemas/user.py
@@ -109,11 +109,13 @@ class UserVerifyErrorResponse(BaseModel):
 
 class UserLoginResponse(BaseModel):
     success: bool
+    last_login_at: datetime | None=None
 
     model_config = {
         "json_schema_extra": {
             "example": {
-                "success": True
+                "success": True,
+                "last_login_at": "2025-09-30T08:13:45.123Z"
             }
         }
     }
